server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "FastAPI backend for browser extension login + project API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.129.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.11.0",
    "python-jose[cryptography]>=3.5.0",
]
